requirements/local.txt
asgiref==3.5.2
Django==4.0.6
django-model-utils==4.0.0
djangorestframework==3.13.1
Pillow==9.2.0
psycopg2-binary==2.9.3
pytz==2019.3
sqlparse==0.4.2
Unipath==1.1
